services/ml-1/requirements.txt
datasets==3.6.0
fastapi==0.115.12
numpy==2.2.6
pandas==2.3.0
pydantic==2.11.5
Requests==2.32.3
seqeval==1.2.2
torch==2.7.0
transformers==4.52.2
uvicorn==0.34.3
python==3.11
google-generativeai==0.8.3